Refactoring tpmmgr - moving some components into a standalone evetpm package - this also helps avoid cyclic import dependency between tpmmgr and zedcloud   i.e. zedcloud depending on tpmmgr for signing and        tpmmgr depending on zedcloud to send HTTP messages
