+++ pkg/pillar/evetpm/tpm.go
@@ -1,0 +1,189 @@
+// Copyright (c) 2020 Zededa, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+package evetpm
+
+import (
+	"crypto"
+	"crypto/ecdsa"
+	"crypto/x509"
+	"encoding/asn1"
+	"encoding/pem"
+	"fmt"
+	"io"
+	"io/ioutil"
+	"math/big"
+	"os"
+	"unsafe"
+
+	"github.com/google/go-tpm/tpm2"
+	"github.com/google/go-tpm/tpmutil"
+	"github.com/lf-edge/eve/pkg/pillar/types"
+	log "github.com/sirupsen/logrus"
+)
+
+const (
+	//TpmDevicePath is the TPM device file path
+	TpmDevicePath = "/dev/tpmrm0"
+
+	//TpmDeviceKeyHdl is the well known TPM permanent handle for device key
+	TpmDeviceKeyHdl tpmutil.Handle = 0x817FFFFF
+
+	//TpmPasswdHdl is the well known TPM NVIndex for TPM Credentials
+	TpmPasswdHdl tpmutil.Handle = 0x1600000
+
+	//TpmCredentialsFileName is the file that holds the dynamically created TPM credentials
+	TpmCredentialsFileName = types.IdentityDirname + "/tpm_credential"
+
+	//MaxPasswdLength is the max length allowed for a TPM password
+	MaxPasswdLength = 7 //limit TPM password to this length
+)
+
+//TpmPrivateKey is Custom implementation of crypto.PrivateKey interface
+type TpmPrivateKey struct {
+	PublicKey crypto.PublicKey
+}
+
+//Helper structure to pack ecdsa signature for ASN1 encoding
+type ecdsaSignature struct {
+	R, S *big.Int
+}
+
+//Public implements crypto.PrivateKey interface
+func (s TpmPrivateKey) Public() crypto.PublicKey {
+	clientCertBytes, err := ioutil.ReadFile(types.DeviceCertName)
+	if err != nil {
+		return nil
+	}
+	block, _ := pem.Decode(clientCertBytes)
+	var cert *x509.Certificate
+	cert, _ = x509.ParseCertificate(block.Bytes)
+	ecdsaPublicKey := cert.PublicKey.(*ecdsa.PublicKey)
+	return ecdsaPublicKey
+}
+
+//Sign implements cryto.PrivateKey interface
+func (s TpmPrivateKey) Sign(r io.Reader, digest []byte, opts crypto.SignerOpts) ([]byte, error) {
+	R, S, err := TpmSign(digest)
+	if err != nil {
+		return nil, err
+	}
+	return asn1.Marshal(ecdsaSignature{R, S})
+}
+
+//ReadOwnerCrdl returns credential specific to this device
+func ReadOwnerCrdl() (string, error) {
+	tpmOwnerPasswdBytes, err := ioutil.ReadFile(TpmCredentialsFileName)
+	if err != nil {
+		return "", err
+	}
+	tpmOwnerPasswd := string(tpmOwnerPasswdBytes)
+	if len(tpmOwnerPasswd) > MaxPasswdLength {
+		tpmOwnerPasswd = tpmOwnerPasswd[0:MaxPasswdLength]
+	}
+	return tpmOwnerPasswd, nil
+}
+
+//TpmSign is used by external packages to get a digest signed by
+//device key in TPM
+func TpmSign(digest []byte) (*big.Int, *big.Int, error) {
+
+	rw, err := tpm2.OpenTPM(TpmDevicePath)
+	if err != nil {
+		return nil, nil, err
+	}
+	defer rw.Close()
+
+	tpmOwnerPasswd, err := ReadOwnerCrdl()
+	if err != nil {
+		log.Fatalf("Error in fetching TPM credentials: %v", err)
+	}
+
+	//XXX This "32" should really come from Hash algo used.
+	if len(digest) > 32 {
+		digest = digest[:32]
+	}
+
+	scheme := &tpm2.SigScheme{
+		Alg:  tpm2.AlgECDSA,
+		Hash: tpm2.AlgSHA256,
+	}
+	sig, err := tpm2.Sign(rw, TpmDeviceKeyHdl,
+		tpmOwnerPasswd, digest, scheme)
+	if err != nil {
+		log.Errorln("Sign using TPM failed")
+		return nil, nil, err
+	}
+	return sig.ECC.R, sig.ECC.S, nil
+}
+
+//FileExists returns true if a file with name filename is found
+func FileExists(filename string) bool {
+	_, err := os.Stat(filename)
+	return err == nil
+}
+
+//IsTpmEnabled checks if TPM is being used by software for creating device cert
+func IsTpmEnabled() bool {
+	return FileExists(types.DeviceCertName) && !FileExists(types.DeviceKeyName)
+}
+
+//GetRandom returns a random []byte of requested length
+func GetRandom(numBytes uint16) ([]byte, error) {
+	rw, err := tpm2.OpenTPM(TpmDevicePath)
+	if err != nil {
+		return nil, err
+	}
+	defer rw.Close()
+	return tpm2.GetRandom(rw, numBytes)
+}
+
+//GetModelName combines vendor1 and vendor2 values into a string
+func GetModelName(vendorValue1 uint32, vendorValue2 uint32) string {
+	uintToByteArr := func(value uint32) []byte {
+		get8 := func(val uint32, offset uint32) uint8 {
+			return (uint8)((val >> ((3 - offset) * 8)) & 0xff)
+		}
+		var i uint32
+		var bytes []byte
+		for i = 0; i < uint32(unsafe.Sizeof(value)); i++ {
+			c := get8(value, i)
+			bytes = append(bytes, c)
+		}
+		return bytes
+	}
+	var model []byte
+	model = append(model, uintToByteArr(vendorValue1)...)
+	model = append(model, uintToByteArr(vendorValue2)...)
+	return string(model)
+}
+
+//GetFirmwareVersion converts v1, v2 values from TPM properties to string
+func GetFirmwareVersion(v1 uint32, v2 uint32) string {
+	get16 := func(val uint32, offset uint32) uint16 {
+		return uint16((val >> ((1 - offset) * 16)) & 0xFFFF)
+	}
+	return fmt.Sprintf("%d.%d.%d.%d", get16(v1, 0), get16(v1, 1),
+		get16(v2, 0), get16(v2, 1))
+}
+
+//GetTpmProperty fetches a given property id, and returns it as uint32
+func GetTpmProperty(propID tpm2.TPMProp) (uint32, error) {
+
+	rw, err := tpm2.OpenTPM(TpmDevicePath)
+	if err != nil {
+		return 0, err
+	}
+	defer rw.Close()
+
+	v, _, err := tpm2.GetCapability(rw, tpm2.CapabilityTPMProperties,
+		1, uint32(propID))
+	if err != nil {
+		return 0, err
+	}
+	prop, ok := v[0].(tpm2.TaggedProperty)
+	if !ok {
+		return 0, fmt.Errorf("Unable to fetch property %d", propID)
+	}
+	return prop.Value, nil
+}
